+++ packages/@aws-cdk-testing/cli-integ/resources/integ.jest.config.js
@@ -20,9 +20,9 @@ module.exports = {
   testTimeout: 600000,
 
   // Affects test.concurrent(), these are self-limiting anyway
-  maxConcurrency: 10,
+  maxWorkers: 50,
   reporters: [
     "default",
-    [ "jest-junit", { suiteName: "jest tests", outputDirectory: "coverage" } ]
+    ["jest-junit", { suiteName: "jest tests", outputDirectory: "coverage" }]
   ]
 };
