+++ packages/@aws-cdk-testing/cli-integ/tests/cli-integ-tests/cdk-migrate/cdk-migrate-deploys-successfully-csharp.integtest.ts
@@ -1,0 +1,13 @@
+import { deploysSuccessfully } from './testcase';
+import { integTest, withCDKMigrateFixture } from "../../../lib";
+
+const language = 'csharp';
+
+integTest(
+  `cdk migrate ${language} deploys successfully`,
+  withCDKMigrateFixture(language, async (fixture) => {
+
+    await deploysSuccessfully(fixture, language);
+
+  }),
+);
